refactor(courses): restructure codebase with major refactor, improve error handling and validation, remove unused error classes, and update import paths
- Refactored the error handling and validation logic across the courses module.
- Removed unused error classes for cleaner code.
- Updated import paths for better maintainability.
- Added comprehensive tests for the refactored code.
- The API remains unchanged.
