+++ backend-proctoring/main/src/config/app.ts
@@ -1,4 +1,4 @@
-import { env } from '@utils/env';
+import { env } from '../utils/env';
 
 function getAppPath() {
   let currentDir = __dirname;
