+++ backend-proctoring/main/src/config/db.ts
@@ -1,4 +1,4 @@
-import { env } from '@utils/env';
+import { env } from '../utils/env';
 
 export const dbConfig = {
     url: env('DB_URL'),
