+++ backend-proctoring/main/src/config/sentry.ts
@@ -1,3 +1,3 @@
-import { env } from "@utils/env";
+import { env } from "utils/env";
 
 export const sentryDSN = env("SENTRY_DSN");
