+++ backend-proctoring/main/src/modules/courses/classes/transformers/CourseVersion.ts
@@ -1,0 +1,52 @@
+import 'reflect-metadata';
+import { Expose, Transform, Type } from "class-transformer";
+import {
+  ObjectIdToString,
+  StringToObjectId,
+} from "shared/constants/transformerConstants";
+import { ICourseVersion } from "shared/interfaces/IUser";
+import { ID } from "shared/types";
+import { Module } from "./Module";
+
+class CourseVersion implements ICourseVersion {
+  @Expose()
+  @Transform(ObjectIdToString.transformer, { toPlainOnly: true })
+  @Transform(StringToObjectId.transformer, { toClassOnly: true })
+  _id?: ID;
+
+  @Expose()
+  @Transform(ObjectIdToString.transformer, { toPlainOnly: true })
+  @Transform(StringToObjectId.transformer, { toClassOnly: true })
+  courseId: ID;
+
+  @Expose()
+  version: string;
+
+  @Expose()
+  description: string;
+
+  @Expose()
+  @Type(() => Module)
+  modules: Module[];
+
+  @Expose()
+  @Type(() => Date)
+  createdAt: Date;
+
+  @Expose()
+  @Type(() => Date)
+  updatedAt: Date;
+
+  constructor(courseVersionPayload?: ICourseVersion) {
+    if (courseVersionPayload) {
+      this.courseId = courseVersionPayload.courseId;
+      this.version = courseVersionPayload.version;
+      this.description = courseVersionPayload.description;
+    }
+    this.modules = [];
+    this.createdAt = new Date();
+    this.updatedAt = new Date();
+  }
+}
+
+export { CourseVersion };
