+++ backend-proctoring/main/src/modules/courses/classes/transformers/transformers.ts
@@ -1,0 +1,6 @@
+export * from './Course';
+export * from './CourseVersion';
+export * from './Module';
+export * from './Section';
+export * from './Item';
+
