+++ backend-proctoring/main/src/modules/courses/classes/validators/CourseVersionPayloadValidators.ts
@@ -1,0 +1,27 @@
+import 'reflect-metadata';
+import { IsEmpty, IsNotEmpty, IsString } from "class-validator";
+import { ICourseVersion, IModule } from "shared/interfaces/IUser";
+
+class CreateCourseVersionPayloadValidator implements ICourseVersion {
+  @IsEmpty()
+  courseId: string;
+
+  @IsNotEmpty()
+  @IsString()
+  version: string;
+
+  @IsNotEmpty()
+  @IsString()
+  description: string;
+
+  @IsEmpty()
+  modules: IModule[];
+
+  @IsEmpty()
+  createdAt: Date;
+
+  @IsEmpty()
+  updatedAt: Date;
+}
+
+export { CreateCourseVersionPayloadValidator };
