+++ backend-proctoring/main/src/modules/courses/classes/validators/validators.ts
@@ -1,0 +1,5 @@
+export * from './CoursePayloadValidators';
+export * from './CourseVersionPayloadValidators';
+export * from './ModuleValidators';
+export * from './SectionValidators';
+export * from './ItemValidators';
