--- backend-proctoring/main/src/modules/courses/errors/ItemErrors.ts
@@ -1,27 +1,0 @@
-export class CreateItemError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "CreateItemError";
-  }
-}
-
-export class FetchItemError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "FetchItemError";
-  }
-}
-
-export class UpdateItemError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "UpdateItemError";
-  }
-}
-
-export class DeleteItemError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = "DeleteItemError";
-    }
-}
