--- backend-proctoring/main/src/modules/courses/errors/VersionErros.ts
@@ -1,27 +1,0 @@
-export class CreateVersionError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "CreateVersionError";
-  }
-}
-
-export class FetchVersionError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "FetchVersionError";
-  }
-}
-
-export class UpdateVersionError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "UpdateVersionError";
-  }
-}
-
-export class DeleteVersionError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "DeleteVersionError";
-  }
-}
