+++ backend-proctoring/main/src/shared/database/interfaces/ICourseRepository.ts
@@ -1,16 +1,23 @@
+import { ItemsGroup } from "modules/courses/classes/transformers/transformers";
 import { ICourse, ICourseVersion, ISection } from "shared/interfaces/IUser";
 
 export interface ICourseRepository {
-    create(course: ICourse): Promise<ICourse | null>;
-    read(id: string): Promise<ICourse | null>;
-    update(id: string, course: Partial<ICourse>): Promise<ICourse | null>;
-    delete(id: string): Promise<boolean>;
-    getAll(): Promise<ICourse[]>;
-    createVersion(courseVersion: ICourseVersion): Promise<ICourseVersion | null>;
-    readVersion(versionId: string): Promise<ICourseVersion | null>;
-    updateVersion(versionId: string, courseVersion: ICourseVersion): Promise<ICourseVersion | null>;
+  create(course: ICourse): Promise<ICourse | null>;
+  read(id: string): Promise<ICourse | null>;
+  update(id: string, course: Partial<ICourse>): Promise<ICourse | null>;
+  delete(id: string): Promise<boolean>;
 
-    /** Section-related operations */
-    readSection(sectionId: string): Promise<ISection | null>;
-    updateSection(sectionId: string, section: Partial<ISection>): Promise<ISection | null>;
+  createVersion(courseVersion: ICourseVersion): Promise<ICourseVersion | null>;
+  readVersion(versionId: string): Promise<ICourseVersion | null>;
+  updateVersion(
+    versionId: string,
+    courseVersion: ICourseVersion
+  ): Promise<ICourseVersion | null>;
+
+  createItemsGroup(itemsGroup: ItemsGroup): Promise<ItemsGroup | null>;
+  readItemsGroup(itemsGroupId: string): Promise<ItemsGroup | null>;
+  updateItemsGroup(
+    itemsGroupId: string,
+    itemsGroup: ItemsGroup
+  ): Promise<ItemsGroup | null>;
 }
