+++ backend-proctoring/main/src/shared/errors/errors.ts
@@ -25,3 +25,17 @@ export class DeleteError extends Error {
     this.name = "DeleteError";
   }
 }
+
+export class ValidationError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "ValidationError";
+  }
+}
+
+export class ItemNotFoundError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "NotFoundError";
+  }
+}
