+++ backend-proctoring/main/src/shared/types.ts
@@ -1,0 +1,7 @@
+import { ObjectId } from "mongodb";
+
+type ID = string | ObjectId | null;
+
+export {
+    ID
+}
