+++ backend-proctoring/main/tests/setup.ts
@@ -1,16 +1,16 @@
-import { MongoMemoryServer } from "mongodb-memory-server";
-import mongoose from "mongoose";
-import "reflect-metadata"; // Ensure decorators work
+// import { MongoMemoryServer } from "mongodb-memory-server";
+// import mongoose from "mongoose";
+// import "reflect-metadata"; // Ensure decorators work
 
-let mongoServer: MongoMemoryServer;
+// let mongoServer: MongoMemoryServer;
 
-beforeAll(async () => {
-  mongoServer = await MongoMemoryServer.create();
-  const mongoUri = mongoServer.getUri();
-  await mongoose.connect(mongoUri);
-});
+// beforeAll(async () => {
+//   mongoServer = await MongoMemoryServer.create();
+//   const mongoUri = mongoServer.getUri();
+//   await mongoose.connect(mongoUri);
+// });
 
-afterAll(async () => {
-  await mongoose.disconnect();
-  await mongoServer.stop();
-});
+// afterAll(async () => {
+//   await mongoose.disconnect();
+//   await mongoServer.stop();
+// });
